Add tpope/vim-fugitive for commits etc
I dislike the way Neogit has it's own filetype for commits. It's hard
(for me) to use with other plugins that check for the `gitcommit`
filetype and have them all work together properly.

Keeping Neogit for now only because it allows for unstaging hunks.
Gitsigns will eventually show staged hunks so they can be unstaged:
https://github.com/lewis6991/gitsigns.nvim/pull/535